requirements.txt
python-telegram-bot>=20.7
google-genai>=1.0.0
supabase>=2.3.0
python-dotenv>=1.0.0
streamlit>=1.32.0
pandas>=2.0.0
